HSEARCH-3105 Move interfaces to the types packages
