+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/search/predicate/impl/AbstractMatchPredicateBuilder.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.search.predicate.impl;
 
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldFormatter;
+import org.hibernate.search.v6poc.backend.lucene.types.formatter.impl.LuceneFieldFormatter;
 import org.hibernate.search.v6poc.search.predicate.spi.MatchPredicateBuilder;
