+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/codec/impl/GeoPointFieldCodec.java
@@ -18,7 +18,6 @@ import org.apache.lucene.document.StoredField;
 import org.apache.lucene.index.IndexableField;
 import org.hibernate.search.v6poc.backend.document.model.Store;
 import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldCodec;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
 import org.hibernate.search.v6poc.backend.spatial.GeoPoint;
 import org.hibernate.search.v6poc.backend.spatial.ImmutableGeoPoint;
