+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/codec/impl/LocalDateFieldCodec.java
@@ -26,7 +26,6 @@ import org.apache.lucene.index.IndexableField;
 import org.hibernate.search.v6poc.backend.document.model.Sortable;
 import org.hibernate.search.v6poc.backend.document.model.Store;
 import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldCodec;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
 
 public final class LocalDateFieldCodec implements LuceneFieldCodec<LocalDate> {
