+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/codec/impl/LuceneFieldCodec.java
@@ -4,13 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.lucene.document.model.impl;
+package org.hibernate.search.v6poc.backend.lucene.types.codec.impl;
 
 import java.util.Collections;
 import java.util.Set;
 
 import org.apache.lucene.document.Document;
 import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
+import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
 
 /**
  * @author Guillaume Smet
