+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/predicate/impl/IntegerFieldPredicateBuilderFactory.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.types.predicate.impl;
 
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldPredicateBuilderFactory;
 import org.hibernate.search.v6poc.backend.lucene.types.formatter.impl.IntegerFieldFormatter;
 
 public final class IntegerFieldPredicateBuilderFactory implements LuceneFieldPredicateBuilderFactory {
