+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/sort/impl/LuceneFieldSortContributor.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.lucene.document.model.impl;
+package org.hibernate.search.v6poc.backend.lucene.types.sort.impl;
 
 import org.hibernate.search.v6poc.backend.lucene.search.sort.impl.LuceneSearchSortCollector;
 import org.hibernate.search.v6poc.search.dsl.sort.SortOrder;
