+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/sort/impl/SortMissingValue.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.lucene.document.model.impl;
+package org.hibernate.search.v6poc.backend.lucene.types.sort.impl;
 
-/**
- * @author Guillaume Smet
- */
-public interface LuceneFieldFormatter<T> {
 
-	T format(Object value);
+public enum SortMissingValue {
+
+	MISSING_FIRST,
+	MISSING_LAST
+
 }
